pyproject.toml
[tool.poetry]
name = "backend-template"
version = "1.1.0"
description = ""
authors = ["Wealize"]

[tool.poetry.dependencies]
python = "^3.10"
djangorestframework = "*"
django-model-utils = "*"
gunicorn = "*"
whitenoise = "*"
Django = "^3.2.9"
django-sendgrid-v5 = "*"
dj-database-url = "*"
sentry-sdk = "0.19.5"
django-storages = "*"
boto3 = "*"
newrelic = "*"
celery = {extras = ["redis"], version="*"}
psycopg2-binary = "*"
django-cors-headers = "*"
django-guardian = "^2.4.0"
django-jazzmin = "^2.4.8"
djangorestframework-simplejwt = "^5.0.0"
django-filter = "^21.1"

[tool.poetry.dev-dependencies]
django-debug-toolbar = "*"
coverage = "*"
pylint = "*"
pylint-django = "*"
pylint-celery = "*"
model_bakery = "*"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
